requirements.txt
imageio==2.16.1
matplotlib==2.2.2
numpy==1.22.3
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
scikit-image==0.19.2
scikit-learn==0.19.2
scipy==1.8.0
# torch==1.11.0
torchvision==0.12.0
tqdm==4.24.0
